Only run newrelic installer if the init script doesn't exist

+++ manifests/php/install.pp
@@ -10,6 +10,7 @@ class newrelic::php::install ($license_key) {
         command => "newrelic-install install",
         path => ['/bin', '/usr/bin'],
         require => Package['newrelic-php5'],
-        notify => Service['php-fpm']
+        notify => Service['php-fpm'],
+        unless => 'test -f /etc/init.d/newrelic-daemon'
     }
 }
